[WIP] Rethink deferred reads
Currently we stuff the result of deferred reads into the cache,
forcing the SHAMap code to call into the cache to get the results
later. This has some pros and cons:

Pros: It's a bit simpler. It minimizes the work the I/O threads
have to do to just populating the cache.

Cons: Nothing locks the results into the cache. If the results
get pushed out of cache before we can get them, bad things
happen. Also, we have to do extra work to get the results out of
the cache.

This is a work in progress to change the design as follows:

When an async read is posted, the node store takes a completion
function to call when the read is complete. The descendAsync code
supports this by taking a completion function to call when
the node store completes the read. The getMissingNodes code
supports this by creating a completion function that stashes
the results and avoids calling back into the node store.

If we accept this change, or one like it, we can get rid of
waitReads and its associated logic. The completion function
can signal a condition variable when all reads complete.